failover-demo2/app-server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
pymongo==4.6.1
kafka-python==2.0.2
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
pydantic==2.5.3
tenacity==8.2.3
python-multipart==0.0.6
pydantic-settings>=2.0.0
motor>=2.6.0
